+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Infrastructure
 * Upgrade pyinstaller to 6.11.1
 
+### Fixed
+* Fix an issue with missing `content_md5` in case the file has been uploaded as a large file
+
 ## [0.10.0] - 2025-01-10
 
 ### Added